+++ changelog.md
@@ -1,4 +1,4 @@
-next
+1.2.2.0
 
 * Important bugfix: versions 1.2.0.0 and 1.2.1.0 were incorrectly displaying
   the lower and upper bounds for measured values on HTML reports.